made a few of the requested changes
